Refactor bot entry point into modular orchestrator
